requirements.txt
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
isort==6.0.1
pytest==8.3.5
pytest-django==4.10.0
pytest-mock==3.14.0
ruff==0.9.10
black==25.1.0
whitenoise==6.9.0
drf-yasg==1.21.10
